Move ObjectMember
